requirements.txt
langchain
langchain-openai
langchain-community
openai
instagrapi
pillow
python-dotenv